removed hard coded paths
