--- tests/read_read_sav.py
@@ -1,50 +1,0 @@
-import pyreadstat
-from pandas.testing import assert_frame_equal
-import pandas as pd
-import numpy as np
-
-columns = [
-    'SERIAL', 'AM_PM_NIGHT', 'AGE', 'ANYUNDER16', 'APORTLATDEG',
-    'APORTLATMIN', 'APORTLATSEC', 'APORTLATNS', 'APORTLONDEG',
-    'APORTLONMIN', 'APORTLONDSEC', 'APORTLONEW', 'ARRIVEDEPART',
-    'BABYFARE', 'BEFAF', 'CHILDFARE', 'CHANGECODE', 'COUNTRYVISIT',
-    'CPORTLATDEG', 'CPORTLATMIN', 'CPORTLATSEC', 'CPORTLATNS',
-    'CPORTLONDEG', 'CPORTLONMIN', 'CPORTLONDSEC', 'CPORTLONEW',
-    'INTDATE', 'DAYTYPE', 'DIRECTLEG', 'DVEXPEND', 'DVFARE',
-    'DVLINECODE', 'DVPACKAGE', 'DVPACKCOST', 'DVPERSONS', 'DVPORTCODE',
-    'EXPENDCODE', 'EXPENDITURE', 'FARE', 'FAREK', 'FLOW', 'HAULKEY',
-    'INTENDLOS', 'INTMONTH', 'KIDAGE', 'LOSKEY', 'MAINCONTRA', 'MIGSI',
-    'NATIONALITY', 'NATIONNAME', 'NIGHTS1', 'NIGHTS2', 'NIGHTS3', 'NIGHTS4',
-    'NIGHTS5', 'NIGHTS6', 'NIGHTS7', 'NIGHTS8', 'NUMADULTS', 'NUMDAYS',
-    'NUMNIGHTS', 'NUMPEOPLE', 'PACKAGEHOL', 'PACKAGEHOLUK', 'PERSONS',
-    'PORTROUTE', 'PACKAGE', 'PROUTELATDEG', 'PROUTELATMIN', 'PROUTELATSEC',
-    'PROUTELATNS', 'PROUTELONDEG', 'PROUTELONMIN', 'PROUTELONSEC',
-    'PROUTELONEW', 'PURPOSE', 'QUARTER', 'RESIDENCE', 'RESPNSE',
-    'SEX', 'SHIFTNO', 'SHUTTLE', 'SINGLERETURN', 'TANDTSI', 'TICKETCOST',
-    'TOWNCODE1', 'TOWNCODE2', 'TOWNCODE3', 'TOWNCODE4', 'TOWNCODE5',
-    'TOWNCODE6', 'TOWNCODE7', 'TOWNCODE8', 'TRANSFER', 'UKFOREIGN',
-    'VEHICLE', 'VISITBEGAN', 'WELSHNIGHTS', 'WELSHTOWN', 'FAREKEY',
-    'TYPEINTERVIEW'
-]
-
-
-def test_load_sav():
-    file_name = '/Users/paul/ONS/ips_services/quarter32017v2.sav'
-    df, meta = pyreadstat.read_sav(file_name)
-
-    df.columns = df.columns.str.upper()
-    df = df.filter(columns, axis=1)
-
-    if 'INTDATE' in df.columns:
-        df['INTDATE'] = df['INTDATE'].astype(str).str.rjust(8, '0')
-
-    def convert_col_to_int(data_frame, col):
-        data_frame[col] = data_frame[col].fillna(-1).astype(int).replace('-1', np.nan)
-
-    df.columns = df.columns.str.upper()
-    df['TANDTSI'] = df['TANDTSI'].round(0)
-    [convert_col_to_int(df, x) for x in ['EXPENDITURE', 'DVEXPEND', 'TANDTSI']]
-
-    df = df.sort_values(by='SERIAL')
-
-    df.to_csv("/Users/paul/Desktop/sav.csv")
